Improved Sequence

+++ src/main/java/ch/bfh/unicrypt/helper/sequence/interfaces/Sequence.java
@@ -86,6 +86,6 @@ public interface Sequence<V>
 
 	public Sequence<V> skip(long n);
 
-	public Sequence<? extends ImmutableArray<V>> group(int groupLength);
+	public Sequence<? extends ImmutableArray<V>> group(long groupLength);
 
 }
